[INJIMOB-3558]: Reduced helper classes
