+++ kotlin/injivcrenderer/src/commonTest/kotlin/io/mosip/injivcrenderer/utils/TemplateHelperTest.kt
@@ -7,10 +7,10 @@ import kotlin.test.assertFailsWith
 import io.mosip.injivcrenderer.TestUtils
 import io.mosip.injivcrenderer.exceptions.VcRendererExceptions
 
-class DigestMultibaseHelperTest {
+class TemplateHelperTest {
 
     private val traceId = "test-trace-id"
-    private val helper = DigestMultibaseHelper(traceId)
+    private val helper = TemplateHelper(traceId)
     private val svgSample = """<svg>Email: {{/credentialSubject/email}}, Mobile: {{/credentialSubject/mobile}}</svg>"""
 
     @Test
